requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
PyMuPDF>=1.23.8
faiss-cpu>=1.7.4
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.0
langchain-ollama>=0.0.6
google-api-python-client>=2.108.0
huggingface-hub>=0.19.0
python-pptx>=0.6.21  # For PPTX support
rich==13.7.0 